+++ fabric/src/main/java/me/pandamods/pandalib/fabric/mixin/ServerPlayerGameModeMixin.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright (C) 2024 Oliver Froberg (The Panda Oliver)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  any later version.
+ *
+ * You should have received a copy of the GNU General Public License
+ *  along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package me.pandamods.pandalib.fabric.mixin;
+
+import com.llamalad7.mixinextras.sugar.Local;
+import me.pandamods.pandalib.event.events.common.BlockEvents;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.server.level.ServerPlayerGameMode;
+import net.minecraft.world.level.block.state.BlockState;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(ServerPlayerGameMode.class)
+public class ServerPlayerGameModeMixin {
+	@Shadow protected ServerLevel level;
+
+	@Shadow @Final protected ServerPlayer player;
+
+	@Inject(method = "destroyBlock",
+			at = @At(
+					value = "INVOKE",
+					target = "Lnet/minecraft/world/level/block/state/BlockState;getBlock()Lnet/minecraft/world/level/block/Block;",
+					ordinal = 0
+			), cancellable = true)
+	public void onBlockBreak(BlockPos blockPos, CallbackInfoReturnable<Boolean> cir, @Local BlockState state) {
+		if (BlockEvents.DESTROY.invoker().destroy(this.level, blockPos, state, this.player))
+			cir.setReturnValue(true);
+	}
+}