+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## Changelog 0.5 Beta
 ### Dependency changes
-* Architectury API is no longer needed as a dependency
+* Architectury API is no longer necessary as a dependency
 
 ### Technical changes
 * Restructured some classes
-* Added an Event system together with the Player Events Join and Quit for both client and server.
+* Added an Event system
+  * Added Player Join and Quit events
+  * Added Block Place and Destroy events
 * Added Reload Listener registration
 * Added more platforms to replace Architectury